Extract Config Vars subcommand to primary command
